\name{popsize-package}
\alias{popsize-package}
\alias{popsize}
\docType{package}
\title{

Estimation de la taille de population
}
\description{

Ce packages permet d'estimer le nombre de personnes dans une zonne bien localisée
}
\details{
\tabular{ll}{
Package: \tab popsize\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-11-23\cr
License: \tab Version Essaie\cr
}
An overview of how to use the package, including the most important functions
}
\author{

Maintainer: <armel.soubeiga@yahoo.fr>
Centre Muraz / Centre de Calcul
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
